requirements.txt
uvicorn==0.12.2
fastapi==0.63.0
psycopg2-binary==2.9.1